added dynamic test
