+++ .github/workflows/ci.yml
@@ -4,6 +4,7 @@ on:
     branches:
       - main
       - dev
+      - dev-8
   pull_request:
     branches:
       - main
@@ -19,7 +20,7 @@ jobs:
         node-version: 18
     - run: npm install
     - name: "Build"
-      run: npx tsc
+      run: npm run build
     - name: "Tests"
       run: npm test
   lint:
