+++ test/helper.ts
@@ -1,6 +1,7 @@
 import { join } from "path";
 import protobuf from "protobufjs";
-import { ProtoDefinition } from "../src/GRPCDefinitionTranslator";
+import { GrpcSymbol, ProtoDefinition } from "../src/GRPCDefinitionTranslator";
+import { INamingTransformer } from "../src/INamingTransformer";
 
 export async function loadFromPbjsDefinition(path: string): Promise<ProtoDefinition>;
 export async function loadFromPbjsDefinition(path: string[]): Promise<ProtoDefinition>;
@@ -14,4 +15,20 @@ export async function loadFromPbjsDefinition(filenames: string | string[]): Prom
 	const description = await protobuf.load(filenames, root);
 	const data = ProtoDefinition.FromPbjs(description.toJSON());
 	return data;
+}
+
+export class MockNamingTransformer implements INamingTransformer {
+	private _modifier?: (symbol: GrpcSymbol) => string;
+	
+	constructor(modifier?: (symbol: GrpcSymbol) => string) {
+		this._modifier = modifier;
+	}
+
+	ConvertSymbol(symbol: GrpcSymbol): string {
+		if (this._modifier) {
+			return this._modifier(symbol);
+		}
+		return symbol.name;
+	}
+
 }
